Rename package component files to kebab-case
