+++ packages/react-transition-switch/src/transition-switch-item.tsx
