+++ packages/react-transition-switch/src/transition-switch.tsx
